Refactor: Update AuthContext import path
