+++ frontend/src/contexts/index.ts
@@ -1,0 +1,2 @@
+export * from './AuthContext';
+
